fix: eliminate sign-comparison warnings and improve path handling in preset browser
- Change currentIndex_ from int to size_t to eliminate compiler sign-comparison warnings
- Redirect error output to cerr instead of cout
- Use std::filesystem::path operator/ for safer path construction instead of string concatenation

+++ src/ui/preset_browser.h
@@ -14,7 +14,7 @@ public:
   const std::vector<std::string>& getPresets() const { return presets_; }
 
   // Get current selection index
-  int getCurrentIndex() const { return currentIndex_; }
+  size_t getCurrentIndex() const { return currentIndex_; }
 
   // Navigate presets
   void nextPreset();
@@ -27,7 +27,7 @@ public:
 private:
   std::vector<std::string> presets_;
   std::string basePath_;
-  int currentIndex_;
+  size_t currentIndex_;
 
   bool loadPresetsFromDirectory(const std::filesystem::path& dir);
 };
